+++ src/lib/command/CommandManager.cpp
@@ -18,6 +18,7 @@
 #include "document/DocumentPath.hpp"
 #include "editor/Editor.hpp"
 #include "format/FormatDialog.hpp"
+#include "format/QuickFormat.hpp"
 #include "help/HelpManager.hpp"
 #include "settings/SettingsDialog.hpp"
 #include "sidebar/SideBarManager.hpp"
@@ -40,8 +41,9 @@ wxBEGIN_EVENT_TABLE(CommandManager, wxEvtHandler)
     EVT_MENU(+CommandId::CloseAll,     CommandManager::onCloseAll)
     EVT_MENU(+CommandId::NewWindow,    CommandManager::onNewWindow)
     EVT_MENU(+CommandId::Quit,         CommandManager::onQuit)
-    EVT_MENU(+CommandId::SessionNew,   CommandManager::onSessionNew)
-    EVT_MENU(+CommandId::SessionClose, CommandManager::onSessionClose)
+    EVT_MENU(+CommandId::SessionNew,    CommandManager::onSessionNew)
+    EVT_MENU(+CommandId::SessionSaveAs, CommandManager::onSessionSaveAs)
+    EVT_MENU(+CommandId::SessionClose,  CommandManager::onSessionClose)
     EVT_MENU_RANGE(wxID_FILE1, wxID_FILE9, CommandManager::onFileHistory)
 
     // Edit
@@ -66,6 +68,7 @@ wxBEGIN_EVENT_TABLE(CommandManager, wxEvtHandler)
     // View
     EVT_MENU(+CommandId::Preferences, CommandManager::onSettings)
     EVT_MENU(+CommandId::Format,      CommandManager::onFormat)
+    EVT_MENU(+CommandId::Reformat,    CommandManager::onReformat)
     EVT_MENU(+CommandId::Subs,        CommandManager::onSubs)
     EVT_MENU(+CommandId::Minimap,     CommandManager::onMinimap)
     EVT_MENU(+CommandId::CompilerLog, CommandManager::onCompilerLog)
@@ -116,6 +119,7 @@ CommandManager::CommandManager(Context& ctx)
         CommandEntry { .id = +CommandId::FindNext,         .name="findNext" },
         CommandEntry { .id = +CommandId::FindPrevious,     .name="findPrevious" },
         CommandEntry { .id = +CommandId::Format,           .name="format" },
+        CommandEntry { .id = +CommandId::Reformat,         .name="reformat" },
         CommandEntry { .id = +CommandId::GotoLine,         .name="gotoLine" },
         CommandEntry { .id = +CommandId::Help,             .name="help"  },
         CommandEntry { .id = +CommandId::IndentDecrease,   .name="indentDec" },
@@ -143,6 +147,7 @@ CommandManager::CommandManager(Context& ctx)
         CommandEntry { .id = +CommandId::SelectAll,        .name="selectAll" },
         CommandEntry { .id = +CommandId::SelectLine,       .name="selectLine" },
         CommandEntry { .id = +CommandId::SessionNew,       .name="sessionNew" },
+        CommandEntry { .id = +CommandId::SessionSaveAs,    .name="sessionSaveAs", .enabled = false },
         CommandEntry { .id = +CommandId::SessionClose,     .name="sessionClose", .enabled = false },
         CommandEntry { .id = +CommandId::ShowExitCode,     .name="showExitCode", .kind = wxITEM_CHECK },
         CommandEntry { .id = +CommandId::Subs,             .name="viewSubs" },
@@ -239,6 +244,10 @@ void CommandManager::onQuit(wxCommandEvent&) {
 
 void CommandManager::onSessionNew(wxCommandEvent&) {
     m_ctx.getDocumentManager().newSession();
+}
+
+void CommandManager::onSessionSaveAs(wxCommandEvent&) {
+    m_ctx.getDocumentManager().saveSessionAs();
 }
 
 void CommandManager::onSessionClose(wxCommandEvent&) {
@@ -345,6 +354,12 @@ void CommandManager::onFormat(wxCommandEvent&) {
         FormatDialog dlg(m_ctx.getUIManager().getMainFrame(), m_ctx, doc);
         dlg.create();
         dlg.ShowModal();
+    }
+}
+
+void CommandManager::onReformat(wxCommandEvent&) {
+    if (auto* doc = m_ctx.getDocumentManager().getActive(); doc != nullptr && doc->getEditor() != nullptr) {
+        QuickFormat(m_ctx).run(*doc->getEditor());
     }
 }